Use explicit target "/" for header logo homepage link.
In production, there's no special need to use
`site.baseurl` as the target for the header logo
homepage link, as it's not dynamic.

There's also no need in the case of a local dev
server, provided a host like `localhost` is used.

`site.baseurl` would only be needed in the case
of serving from a gh-pages project site, where URL
paths are offset from the docroot, e.g., /webrtc-org

We don't care about that case. Fixes #15.