Updating "Component Builds" documentation
Component builds are not supported by the WebRTC build.
This pull request fixes the documentation with a notice in all the
three pages under 'Native Code'.

Not supporting component builds means that there is no difference
between is_component_build=true and is_component_build=false.

To support this Gn argument we should update all the BUILD.gn files to use
the target type 'component' instead of 'rtc_static_library'/'rtc_shared_library'.
This would give Gn the power to generate shared libraries or static libraries
looking at the value of 'is_component_build'.

This is a non trivial effort and we don't think that the generation
of single WebRTC shared "sub-libraries" is common enough to justify the effort.

These targets are still available on their own but their type is statically
declared in the BUILD.gn file.